Serialize first-class version of functions
Summary: updates the way we serialize to write out the first-class
version of the function. This allows the removal of hacks in the
import pathway for the module hooks. This also updates the way forks
are serialized such that they are no longer required to be stand alone
functions/methods at the top level. Instead they are serialized as
inner functions and enough compiler infrastructure is present to allow
fork to invoke one of those inner functions on reload

gh-metadata: pytorch pytorch 19723 gh/zdevito/30/head